server: use search endpoint to increase packetsize
Since searche endpoint expects a Pokémon struct, the size of the payload remains 292.
However, you can send multiple Pokémon structs, the expected amount is given in the body of the request.

We keep the regular result endpoint since this is the initial entrypoint. Preferably keep it this way
so we may edit the code related to the search endpoint on the DS side without having to invalidate the cache.

Signed-off-by: Jorik Devreese <jorikdevreese@hotmail.be>